filters: centralize boolean query parsing and dedupe view helpers
Several apps had copied parse_bool/_parse_bool/_truthy helpers with
slightly different implementations.

Add a shared parse_bool + is_truthy utility to makerfex_backend.filters
and update viewsets to import the shared helper instead of maintaining
local copies.

Also normalize Customers vip filtering to accept common truthy values
(vip=true/yes/on) while preserving vip=1 compatibility.

No API breaking changes; improves consistency across server-driven
tables and reduces drift across apps.

Refs #9
